+++ src/Pages/Account/Account.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Account.css';
 
-export default function Account() {
+export function Account() {
   return (
     <>
     <div className="notes-container">
