+++ src/Pages/index.js
@@ -1,0 +1,8 @@
+export { Account } from './Account/Account';
+export { Archive } from './archivePage/archive';
+export { HomePage } from './Homepage/HomePage';
+export { Label } from './Label/Label';
+export { Login } from './Login/Login';
+export { NotesPage } from './NotesPage/NotesPage';
+export { SignUp } from './SignUp/SignUp';
+export { Trash } from './Trash/trash';
