+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { Navbar } from './Navbar/Navbar';
+export { RequiresAuth } from './RequiresAuth/RequiresAuth';
+export { Sidebar } from './Sidebar/Sidebar';
+export { toast } from './toast/toast';
