+++ src/components/toast/toast.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './toast.css';
 
 
-export default function toast({value , color}) {
+export function toast({value , color}) {
   return (
     <>
      <div className='alert' style={{backgroundColor: color}}>{value} <span className='close-btn'>X</span></div>
