+++ src/context/index.js
@@ -1,0 +1,4 @@
+export { useTrashContext, TrashContextProvider } from './trashContext';
+export { useNote , NoteContextProvider } from './note-context';
+export { useArchiveNote , ArchiveContextProvider } from './archive-note-context';
+export { useToastContext, ToastContextProvider } from  './toastContext';
