+++ src/index.js
@@ -4,10 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from 'react-router-dom';
-import { NoteContextProvider } from "./context/note-context";
-import { ArchiveContextProvider } from "./context/archive-note-context";
-import { ToastContextProvider } from "./context/toastContext";
-import { TrashContextProvider } from "./context/trashContext";
+import { NoteContextProvider, TrashContextProvider, ToastContextProvider, ArchiveContextProvider } from "./context/index";
 
 // Call make Server
 makeServer();
